pyproject.toml
[project]
name = "payments-pipeline"
version = "0.3.1"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9",
    "pandas>=2.1",
]

[dependency-groups]
dev = ["pytest>=7.0", "pytest-mock>=3.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
